Added the auth provider and service layer.
