+++ mynotes/lib/main.dart
@@ -1,11 +1,9 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'constants/routes.dart';
-import 'firebase_options.dart';
-import 'views/login_view.dart';
-import 'dart:developer' as devtools show log;
 
+import 'services/auth/auth_service.dart';
+import 'constants/routes.dart';
+import 'views/login_view.dart';
+import 'views/notes_view.dart';
 import 'views/register_view.dart';
 import 'views/verify_email_view.dart';
 
@@ -33,16 +31,16 @@ class Homepage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final authService = AuthService.firebase();
+
     return FutureBuilder(
-      future: Firebase.initializeApp(
-        options: DefaultFirebaseOptions.currentPlatform,
-      ),
+      future: authService.initialize(),
       builder: (BuildContext context, snapshot) {
         switch (snapshot.connectionState) {
           case ConnectionState.done:
-            final user = FirebaseAuth.instance.currentUser;
+            final user = authService.currentUser;
             if (user != null) {
-              if (user.emailVerified) {
+              if (user.isEmailedVerified) {
                 return const NotesView();
               } else {
                 return const VerifyEmailView();
@@ -56,80 +54,4 @@ class Homepage extends StatelessWidget {
       },
     );
   }
-}
-
-enum MenuAction { logout }
-
-class NotesView extends StatefulWidget {
-  const NotesView({Key? key}) : super(key: key);
-
-  @override
-  State<NotesView> createState() => _NotesViewState();
-}
-
-class _NotesViewState extends State<NotesView> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Main UI'),
-        actions: [
-          PopupMenuButton<MenuAction>(
-            onSelected: (value) async {
-              devtools.log(value.toString());
-
-              switch (value) {
-                case MenuAction.logout:
-                  final shouldLogout = await showLogoutDialog(context);
-                  devtools.log(shouldLogout.toString());
-
-                  if (shouldLogout) {
-                    await FirebaseAuth.instance.signOut();
-
-                    Navigator.of(context).pushNamedAndRemoveUntil(
-                      loginRoute,
-                      (route) => false,
-                    );
-                  }
-                  break;
-              }
-            },
-            itemBuilder: (context) {
-              return const [
-                PopupMenuItem<MenuAction>(
-                  value: MenuAction.logout,
-                  child: Text('Logout'),
-                )
-              ];
-            },
-          ),
-        ],
-      ),
-      body: const Text('Hello World'),
-    );
-  }
-}
-
-Future<bool> showLogoutDialog(BuildContext context) {
-  return showDialog<bool>(
-    context: context,
-    builder: (context) {
-      return AlertDialog(
-        title: const Text('Sign out'),
-        content: const Text('Are you sure you want to sign out?'),
-        actions: [
-          TextButton(
-              onPressed: () {
-                Navigator.of(context).pop(false);
-              },
-              child: const Text('Cancel')),
-          TextButton(
-              onPressed: () {
-                Navigator.of(context).pop(true);
-              },
-              child: const Text('Logout'))
-        ],
-      );
-    },
-  ).then((value) => value ?? false);
 }
