+++ mynotes/lib/services/auth/auth_provider.dart
@@ -1,6 +1,8 @@
 import 'auth_user.dart';
 
 abstract class AuthProvider {
+  Future<void> initialize();
+
   AuthUser? get currentUser;
 
   Future<AuthUser> logIn({
