+++ mynotes/lib/services/auth/firebase_auth_provider.dart
@@ -1,5 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart'
     show FirebaseAuth, FirebaseAuthException;
+import 'package:firebase_core/firebase_core.dart';
+import '../../firebase_options.dart';
 import 'auth_provider.dart';
 import 'auth_user.dart';
 import 'auth_exceptions.dart';
@@ -91,5 +93,12 @@ class FirebaseAuthProvider implements AuthProvider {
     } else {
       throw UserNotFoundAuthException();
     }
+  }
+
+  @override
+  Future<void> initialize() async {
+    await Firebase.initializeApp(
+      options: DefaultFirebaseOptions.currentPlatform,
+    );
   }
 }
