+++ mynotes/lib/views/notes_view.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/material.dart';
+import 'dart:developer' as devtools show log;
+
+import '../constants/routes.dart';
+import '../enums/menu_action.dart';
+import '../services/auth/auth_service.dart';
+
+class NotesView extends StatefulWidget {
+  const NotesView({Key? key}) : super(key: key);
+
+  @override
+  State<NotesView> createState() => _NotesViewState();
+}
+
+class _NotesViewState extends State<NotesView> {
+  final authService = AuthService.firebase();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Main UI'),
+        actions: [
+          PopupMenuButton<MenuAction>(
+            onSelected: (value) async {
+              devtools.log(value.toString());
+
+              switch (value) {
+                case MenuAction.logout:
+                  final shouldLogout = await showLogoutDialog(context);
+                  devtools.log(shouldLogout.toString());
+
+                  if (shouldLogout) {
+                    await authService.logOut();
+
+                    Navigator.of(context).pushNamedAndRemoveUntil(
+                      loginRoute,
+                          (route) => false,
+                    );
+                  }
+                  break;
+              }
+            },
+            itemBuilder: (context) {
+              return const [
+                PopupMenuItem<MenuAction>(
+                  value: MenuAction.logout,
+                  child: Text('Logout'),
+                )
+              ];
+            },
+          ),
+        ],
+      ),
+      body: const Text('Hello World'),
+    );
+  }
+}
+
+Future<bool> showLogoutDialog(BuildContext context) {
+  return showDialog<bool>(
+    context: context,
+    builder: (context) {
+      return AlertDialog(
+        title: const Text('Sign out'),
+        content: const Text('Are you sure you want to sign out?'),
+        actions: [
+          TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(false);
+              },
+              child: const Text('Cancel')),
+          TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(true);
+              },
+              child: const Text('Logout'))
+        ],
+      );
+    },
+  ).then((value) => value ?? false);
+}
